去掉concat的再量化部分(因为新的模型不需要再量化requant)

+++ CSIM/Sparse_HLS/Sparse_HLS/Sparse_HLS.prj/sol/.autopilot/db/sparse.pp.0.cpp
@@ -70312,7 +70312,7 @@ void quant(hls::stream<typename WideType<t_Quant_DataType, nPE>::t_TypeInt> &dat
    unsigned int fm_COLS,
    hls::stream<typename WideType<t_DataType_OUT, nPE>::t_TypeInt> &requant_stream_out,
    int quant_shift,
-   int quant_mul
+   int16_t quant_mul
    ){
 
     int64_t ONE = static_cast<int64_t> (1);
@@ -70321,7 +70321,7 @@ void quant(hls::stream<typename WideType<t_Quant_DataType, nPE>::t_TypeInt> &dat
   WideType<t_Quant_DataType, nPE> dataValue = data_stream_out.read();
   WideType<t_DataType_OUT, nPE> outValue;
   VITIS_LOOP_103_2: for(int j = 0; j < nPE; j++){
-   int64_t temp = static_cast<int64_t>(dataValue[j]);
+   int32_t temp = static_cast<int32_t>(dataValue[j]);
    int right_shift = quant_shift > 0 ? quant_shift : 0;
       int left_shift = quant_shift > 0 ? 0 : (-quant_shift);
       if (left_shift > 0){
@@ -70329,7 +70329,7 @@ void quant(hls::stream<typename WideType<t_Quant_DataType, nPE>::t_TypeInt> &dat
       }
       temp = temp * quant_mul;
       int total_right_shift = right_shift + 31;
-      int64_t pos_rounding_value = (ONE << (total_right_shift - ONE));
+      int32_t pos_rounding_value = (ONE << (total_right_shift - ONE));
       temp = temp + pos_rounding_value;
       temp = temp >> total_right_shift;
 
@@ -70389,7 +70389,7 @@ __attribute__((sdx_kernel("sparse", 0))) void sparse(
  ap_uint<256> *inputs,
  ap_uint<256> *outputs,
  int shift,
- int mul,
+ int16_t mul,
  bool &sparse_flag);
 # 2 "/home/ytq/codeField/Prediction_Model_Accelerator/CSIM/Sparse_HLS/Sparse_HLS/src/sparse.cpp" 2
 
@@ -70404,7 +70404,7 @@ __attribute__((sdx_kernel("sparse", 0))) void sparse(
  ap_uint<256> *inputs,
  ap_uint<256> *outputs,
  int quant_shift,
- int quant_mul,
+ int16_t quant_mul,
  bool &sparse_flag)
 {
 #line 55 "/home/ytq/codeField/Prediction_Model_Accelerator/CSIM/Sparse_HLS/Sparse_HLS/run_hls.tcl"
